pyproject.toml
[tool.poetry]
name = "test-taska"
version = "1.0.2"
description = ""
authors = ["DanyaSh <danilashundeev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
aiogram = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
